requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.9.0
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.2.1
colorama==0.4.6
fastapi==0.115.12
frozenlist==1.4.1
h11==0.16.0
httptools==0.6.4
idna==3.7
multidict==6.0.5
openai==0.28.0
pydantic==2.11.5
pydantic_core==2.33.2
python-dotenv==1.1.0
python-multipart==0.0.20
PyYAML==6.0.2
requests==2.31.0
sniffio==1.3.1
starlette==0.46.2
tqdm==4.66.2
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.2.1
uvicorn==0.34.3
watchfiles==1.0.5
websockets==15.0.1
yarl==1.9.4
